Restructure into one file per changeset action
- Merge deploy_cache.go and deploy_proxy.go into deploy.go
- Rename configure_cache, remove_feeds, and freeze_feeds to match
  their changeset names (set_feed_configs, remove_feed_configs,
  set_feed_frozen)
- Split changeset_test.go into per-action test files, mirroring the
  Aptos suite; shared fakes, env builder, and datastore seeding move
  to testutil_test.go
